Factor damped wave gauge functions for reuse
