+++ src/Evolution/Systems/GeneralizedHarmonic/GaugeSourceFunctions/DampedHarmonic.cpp
@@ -12,6 +12,7 @@
 #include "DataStructures/Tensor/EagerMath/DotProduct.hpp"
 #include "DataStructures/Tensor/Tensor.hpp"
 #include "DataStructures/Variables.hpp"
+#include "Evolution/Systems/GeneralizedHarmonic/GaugeSourceFunctions/DampedWaveHelpers.hpp"
 #include "PointwiseFunctions/GeneralRelativity/ComputeGhQuantities.hpp"
 #include "PointwiseFunctions/GeneralRelativity/ComputeSpacetimeQuantities.hpp"
 #include "Utilities/ConstantExpressions.hpp"
@@ -22,49 +23,8 @@
 
 // IWYU pragma: no_forward_declare Tensor
 
-namespace GeneralizedHarmonic {
-namespace gauges {
+namespace GeneralizedHarmonic::gauges {
 namespace DampedHarmonicGauge_detail {
-// Spatial weight function used in the damped harmonic gauge source
-// function.
-//
-// The spatial weight function is: \f[ W(x^i) = \exp(- (r / \sigma_r)^2) \f]
-// This function can be written with an extra factor inside the exponent in
-// literature, e.g. \cite Deppe2018uye. We absorb that in \f$\sigma_r\f$.
-template <size_t SpatialDim, typename Frame, typename DataType>
-void spatial_weight_function(const gsl::not_null<Scalar<DataType>*> weight,
-                             const tnsr::I<DataType, SpatialDim, Frame>& coords,
-                             const double sigma_r) noexcept {
-  if (UNLIKELY(get_size(get(*weight)) != get_size(get<0>(coords)))) {
-    *weight = Scalar<DataType>(get_size(get<0>(coords)));
-  }
-  const auto r_squared = dot_product(coords, coords);
-  get(*weight) = exp(-get(r_squared) / square(sigma_r));
-}
-
-// Spacetime derivatives of the spatial weight function that enters the
-// damped harmonic gauge source function.
-//
-// Compute the derivatives :
-// \partial_a W(x^i)= \partial_a \exp(- (r/\sigma_r)^2)
-//                  = (-2 * x^i / \sigma_r^2) * exp(-(r/\sigma_r)^2)
-template <size_t SpatialDim, typename Frame, typename DataType>
-void spacetime_deriv_of_spatial_weight_function(
-    const gsl::not_null<tnsr::a<DataType, SpatialDim, Frame>*> d4_weight,
-    const tnsr::I<DataType, SpatialDim, Frame>& coords, const double sigma_r,
-    const Scalar<DataType>& spatial_weight_function) noexcept {
-  if (UNLIKELY(get_size(get<0>(*d4_weight)) != get_size(get<0>(coords)))) {
-    *d4_weight = tnsr::a<DataType, SpatialDim, Frame>(get_size(get<0>(coords)));
-  }
-  // use 0th component to avoid allocations
-  get<0>(*d4_weight) = get(spatial_weight_function) * (-2. / square(sigma_r));
-  for (size_t i = 0; i < SpatialDim; ++i) {
-    d4_weight->get(1 + i) = get<0>(*d4_weight) * coords.get(i);
-  }
-  // time derivative of weight function is zero
-  get<0>(*d4_weight) = 0.;
-}
-
 // Roll-on function for the damped harmonic gauge.
 //
 // For times after \f$t_0\f$, compute the roll-on function
@@ -91,195 +51,6 @@ double time_deriv_of_roll_on_function(const double time, const double t_start,
   const double time_since_start_over_sigma = (time - t_start) / sigma_t;
   return exp(-pow<4>(time_since_start_over_sigma)) * 4. *
          pow<3>(time_since_start_over_sigma) / sigma_t;
-}
-
-// The log factor that appears in damped harmonic gauge source function.
-//
-// Calculates:  \f$ logF = \mathrm{log}(g^p/N) \f$.
-template <typename DataType>
-void log_factor_metric_lapse(const gsl::not_null<Scalar<DataType>*> logfac,
-                             const Scalar<DataType>& lapse,
-                             const Scalar<DataType>& sqrt_det_spatial_metric,
-                             const double exponent) noexcept {
-  if (UNLIKELY(get_size(get(*logfac)) != get_size(get(lapse)))) {
-    *logfac = Scalar<DataType>(get_size(get(lapse)));
-  }
-  // branching below is to avoid using pow for performance reasons
-  if (exponent == 0.) {
-    get(*logfac) = -log(get(lapse));
-  } else if (exponent == 0.5) {
-    get(*logfac) = log(get(sqrt_det_spatial_metric) / get(lapse));
-  } else {
-    get(*logfac) =
-        2. * exponent * log(get(sqrt_det_spatial_metric)) - log(get(lapse));
-  }
-}
-
-template <typename DataType>
-Scalar<DataType> log_factor_metric_lapse(
-    const Scalar<DataType>& lapse,
-    const Scalar<DataType>& sqrt_det_spatial_metric,
-    const double exponent) noexcept {
-  Scalar<DataType> logfac{};
-  log_factor_metric_lapse(make_not_null(&logfac), lapse,
-                          sqrt_det_spatial_metric, exponent);
-  return logfac;
-}
-
-// Spacetime derivatives of the log factor that appears in the
-// damped harmonic gauge source function.
-//
-// Computes the spacetime derivatives:
-//  \partial_a logF = (p/g)\partial_a g - (1/N)\partial_a N
-template <size_t SpatialDim, typename Frame, typename DataType>
-void spacetime_deriv_of_log_factor_metric_lapse(
-    const gsl::not_null<tnsr::a<DataType, SpatialDim, Frame>*> d4_logfac,
-    const Scalar<DataType>& lapse,
-    const tnsr::I<DataType, SpatialDim, Frame>& shift,
-    const tnsr::A<DataType, SpatialDim, Frame>& spacetime_unit_normal,
-    const tnsr::II<DataType, SpatialDim, Frame>& inverse_spatial_metric,
-    const Scalar<DataType>& sqrt_det_spatial_metric,
-    const tnsr::ii<DataType, SpatialDim, Frame>& dt_spatial_metric,
-    const tnsr::aa<DataType, SpatialDim, Frame>& pi,
-    const tnsr::iaa<DataType, SpatialDim, Frame>& phi,
-    const double exponent) noexcept {
-  if (UNLIKELY(get_size(get<0>(*d4_logfac)) != get_size(get(lapse)))) {
-    *d4_logfac = tnsr::a<DataType, SpatialDim, Frame>(get_size(get(lapse)));
-  }
-  // Use a TempBuffer to reduce total number of allocations. This is especially
-  // important in a multithreaded environment.
-  TempBuffer<tmpl::list<::Tags::Tempa<0, SpatialDim, Frame, DataType>,
-                        ::Tags::Tempi<1, SpatialDim, Frame, DataType>,
-                        ::Tags::Tempa<2, SpatialDim, Frame, DataType>,
-                        ::Tags::TempScalar<3, DataType>,
-                        ::Tags::TempScalar<4, DataType>>>
-      buffer(get_size(get(lapse)));
-  auto& d_g = get<::Tags::Tempa<0, SpatialDim, Frame, DataType>>(buffer);
-  auto& d3_lapse = get<::Tags::Tempi<1, SpatialDim, Frame, DataType>>(buffer);
-  auto& d_lapse = get<::Tags::Tempa<2, SpatialDim, Frame, DataType>>(buffer);
-  auto& dt_lapse = get<::Tags::TempScalar<3, DataType>>(buffer);
-  auto& one_over_lapse = get<::Tags::TempScalar<4, DataType>>(buffer);
-
-  // Get \f$ \partial_a g\f$
-  spacetime_deriv_of_det_spatial_metric<SpatialDim, Frame, DataType>(
-      make_not_null(&d_g), sqrt_det_spatial_metric, inverse_spatial_metric,
-      dt_spatial_metric, phi);
-  // Get \f$ \partial_a N\f$
-  time_deriv_of_lapse<SpatialDim, Frame, DataType>(
-      make_not_null(&dt_lapse), lapse, shift, spacetime_unit_normal, phi, pi);
-  spatial_deriv_of_lapse<SpatialDim, Frame, DataType>(
-      make_not_null(&d3_lapse), lapse, spacetime_unit_normal, phi);
-  get<0>(d_lapse) = get(dt_lapse);
-  for (size_t i = 0; i < SpatialDim; ++i) {
-    d_lapse.get(1 + i) = d3_lapse.get(i);
-  }
-  // Compute
-  get(one_over_lapse) = 1. / get(lapse);
-  if (exponent == 0.) {
-    for (size_t a = 0; a < SpatialDim + 1; ++a) {
-      d4_logfac->get(a) = -get(one_over_lapse) * d_lapse.get(a);
-    }
-  } else {
-    const auto p_over_g = exponent / square(get(sqrt_det_spatial_metric));
-    for (size_t a = 0; a < SpatialDim + 1; ++a) {
-      d4_logfac->get(a) =
-          p_over_g * d_g.get(a) - get(one_over_lapse) * d_lapse.get(a);
-    }
-  }
-}
-
-template <size_t SpatialDim, typename Frame, typename DataType>
-tnsr::a<DataType, SpatialDim, Frame> spacetime_deriv_of_log_factor_metric_lapse(
-    const Scalar<DataType>& lapse,
-    const tnsr::I<DataType, SpatialDim, Frame>& shift,
-    const tnsr::A<DataType, SpatialDim, Frame>& spacetime_unit_normal,
-    const tnsr::II<DataType, SpatialDim, Frame>& inverse_spatial_metric,
-    const Scalar<DataType>& sqrt_det_spatial_metric,
-    const tnsr::ii<DataType, SpatialDim, Frame>& dt_spatial_metric,
-    const tnsr::aa<DataType, SpatialDim, Frame>& pi,
-    const tnsr::iaa<DataType, SpatialDim, Frame>& phi,
-    const double exponent) noexcept {
-  tnsr::a<DataType, SpatialDim, Frame> d4_logfac{};
-  spacetime_deriv_of_log_factor_metric_lapse(
-      make_not_null(&d4_logfac), lapse, shift, spacetime_unit_normal,
-      inverse_spatial_metric, sqrt_det_spatial_metric, dt_spatial_metric, pi,
-      phi, exponent);
-  return d4_logfac;
-}
-
-// Spacetime derivatives of the log factor (that appears in
-// damped harmonic gauge source function), raised to an exponent.
-//
-// Computes the spacetime derivatives:
-//  \partial_a (logF)^q = q (logF)^{q-1} \partial_a (logF)
-template <size_t SpatialDim, typename Frame, typename DataType>
-void spacetime_deriv_of_power_log_factor_metric_lapse(
-    const gsl::not_null<tnsr::a<DataType, SpatialDim, Frame>*> d4_powlogfac,
-    const Scalar<DataType>& lapse,
-    const tnsr::I<DataType, SpatialDim, Frame>& shift,
-    const tnsr::A<DataType, SpatialDim, Frame>& spacetime_unit_normal,
-    const tnsr::II<DataType, SpatialDim, Frame>& inverse_spatial_metric,
-    const Scalar<DataType>& sqrt_det_spatial_metric,
-    const tnsr::ii<DataType, SpatialDim, Frame>& dt_spatial_metric,
-    const tnsr::aa<DataType, SpatialDim, Frame>& pi,
-    const tnsr::iaa<DataType, SpatialDim, Frame>& phi, const double g_exponent,
-    const int exponent) noexcept {
-  if (UNLIKELY(get_size(get<0>(*d4_powlogfac)) != get_size(get(lapse)))) {
-    *d4_powlogfac = tnsr::a<DataType, SpatialDim, Frame>(get_size(get(lapse)));
-  }
-  // Use a TempBuffer to reduce total number of allocations. This is especially
-  // important in a multithreaded environment.
-  TempBuffer<tmpl::list<::Tags::Tempa<0, SpatialDim, Frame, DataType>,
-                        ::Tags::TempScalar<1, DataType>,
-                        ::Tags::TempScalar<2, DataType>>>
-      buffer(get_size(get(lapse)));
-  auto& dlogfac = get<::Tags::Tempa<0, SpatialDim, Frame, DataType>>(buffer);
-  auto& logfac = get<::Tags::TempScalar<1, DataType>>(buffer);
-  auto& prefac = get<::Tags::TempScalar<2, DataType>>(buffer);
-
-  // Compute derivative
-  spacetime_deriv_of_log_factor_metric_lapse<SpatialDim, Frame, DataType>(
-      make_not_null(&dlogfac), lapse, shift, spacetime_unit_normal,
-      inverse_spatial_metric, sqrt_det_spatial_metric, dt_spatial_metric, pi,
-      phi, g_exponent);
-  // Apply pre-factor
-  if (UNLIKELY(exponent == 0)) {
-    for (size_t a = 0; a < SpatialDim + 1; ++a) {
-      d4_powlogfac->get(a) = 0.;
-    }
-  } else if (UNLIKELY(exponent == 1)) {
-    for (size_t a = 0; a < SpatialDim + 1; ++a) {
-      d4_powlogfac->get(a) = dlogfac.get(a);
-    }
-  } else {
-    log_factor_metric_lapse<DataType>(make_not_null(&logfac), lapse,
-                                      sqrt_det_spatial_metric, g_exponent);
-    get(prefac) =
-        static_cast<double>(exponent) * pow(get(logfac), exponent - 1);
-    for (size_t a = 0; a < SpatialDim + 1; ++a) {
-      d4_powlogfac->get(a) = get(prefac) * dlogfac.get(a);
-    }
-  }
-}
-
-template <size_t SpatialDim, typename Frame, typename DataType>
-tnsr::a<DataType, SpatialDim, Frame>
-spacetime_deriv_of_power_log_factor_metric_lapse(
-    const Scalar<DataType>& lapse,
-    const tnsr::I<DataType, SpatialDim, Frame>& shift,
-    const tnsr::A<DataType, SpatialDim, Frame>& spacetime_unit_normal,
-    const tnsr::II<DataType, SpatialDim, Frame>& inverse_spatial_metric,
-    const Scalar<DataType>& sqrt_det_spatial_metric,
-    const tnsr::ii<DataType, SpatialDim, Frame>& dt_spatial_metric,
-    const tnsr::aa<DataType, SpatialDim, Frame>& pi,
-    const tnsr::iaa<DataType, SpatialDim, Frame>& phi, const double g_exponent,
-    const int exponent) noexcept {
-  tnsr::a<DataType, SpatialDim, Frame> d4_powlogfac{};
-  spacetime_deriv_of_power_log_factor_metric_lapse(
-      make_not_null(&d4_powlogfac), lapse, shift, spacetime_unit_normal,
-      inverse_spatial_metric, sqrt_det_spatial_metric, dt_spatial_metric, pi,
-      phi, g_exponent, exponent);
-  return d4_powlogfac;
 }
 }  // namespace DampedHarmonicGauge_detail
 
@@ -710,78 +481,6 @@ void spacetime_deriv_damped_harmonic_h(
 #define FRAME(data) BOOST_PP_TUPLE_ELEM(2, data)
 #define DTYPE_SCAL(data) BOOST_PP_TUPLE_ELEM(0, data)
 
-#define INSTANTIATE(_, data)                                                  \
-  template void DampedHarmonicGauge_detail::spatial_weight_function(          \
-      const gsl::not_null<Scalar<DTYPE(data)>*> weight,                       \
-      const tnsr::I<DTYPE(data), DIM(data), FRAME(data)>& coords,             \
-      const double sigma_r) noexcept;                                         \
-  template void                                                               \
-  DampedHarmonicGauge_detail::spacetime_deriv_of_spatial_weight_function(     \
-      const gsl::not_null<tnsr::a<DTYPE(data), DIM(data), FRAME(data)>*>      \
-          d4_weight,                                                          \
-      const tnsr::I<DTYPE(data), DIM(data), FRAME(data)>& coords,             \
-      const double sigma_r,                                                   \
-      const Scalar<DTYPE(data)>& weight_function) noexcept;                   \
-  template void                                                               \
-  DampedHarmonicGauge_detail::spacetime_deriv_of_log_factor_metric_lapse(     \
-      const gsl::not_null<tnsr::a<DTYPE(data), DIM(data), FRAME(data)>*>      \
-          d4_logfac,                                                          \
-      const Scalar<DTYPE(data)>& lapse,                                       \
-      const tnsr::I<DTYPE(data), DIM(data), FRAME(data)>& shift,              \
-      const tnsr::A<DTYPE(data), DIM(data), FRAME(data)>&                     \
-          spacetime_unit_normal,                                              \
-      const tnsr::II<DTYPE(data), DIM(data), FRAME(data)>&                    \
-          inverse_spatial_metric,                                             \
-      const Scalar<DTYPE(data)>& sqrt_det_spatial_metric,                     \
-      const tnsr::ii<DTYPE(data), DIM(data), FRAME(data)>& dt_spatial_metric, \
-      const tnsr::aa<DTYPE(data), DIM(data), FRAME(data)>& pi,                \
-      const tnsr::iaa<DTYPE(data), DIM(data), FRAME(data)>& phi,              \
-      const double exponent) noexcept;                                        \
-  template tnsr::a<DTYPE(data), DIM(data), FRAME(data)>                       \
-  DampedHarmonicGauge_detail::spacetime_deriv_of_log_factor_metric_lapse(     \
-      const Scalar<DTYPE(data)>& lapse,                                       \
-      const tnsr::I<DTYPE(data), DIM(data), FRAME(data)>& shift,              \
-      const tnsr::A<DTYPE(data), DIM(data), FRAME(data)>&                     \
-          spacetime_unit_normal,                                              \
-      const tnsr::II<DTYPE(data), DIM(data), FRAME(data)>&                    \
-          inverse_spatial_metric,                                             \
-      const Scalar<DTYPE(data)>& sqrt_det_spatial_metric,                     \
-      const tnsr::ii<DTYPE(data), DIM(data), FRAME(data)>& dt_spatial_metric, \
-      const tnsr::aa<DTYPE(data), DIM(data), FRAME(data)>& pi,                \
-      const tnsr::iaa<DTYPE(data), DIM(data), FRAME(data)>& phi,              \
-      const double exponent) noexcept;                                        \
-  template void DampedHarmonicGauge_detail::                                  \
-      spacetime_deriv_of_power_log_factor_metric_lapse(                       \
-          const gsl::not_null<tnsr::a<DTYPE(data), DIM(data), FRAME(data)>*>  \
-              d4_powlogfac,                                                   \
-          const Scalar<DTYPE(data)>& lapse,                                   \
-          const tnsr::I<DTYPE(data), DIM(data), FRAME(data)>& shift,          \
-          const tnsr::A<DTYPE(data), DIM(data), FRAME(data)>&                 \
-              spacetime_unit_normal,                                          \
-          const tnsr::II<DTYPE(data), DIM(data), FRAME(data)>&                \
-              inverse_spatial_metric,                                         \
-          const Scalar<DTYPE(data)>& sqrt_det_spatial_metric,                 \
-          const tnsr::ii<DTYPE(data), DIM(data), FRAME(data)>&                \
-              dt_spatial_metric,                                              \
-          const tnsr::aa<DTYPE(data), DIM(data), FRAME(data)>& pi,            \
-          const tnsr::iaa<DTYPE(data), DIM(data), FRAME(data)>& phi,          \
-          const double g_exponent, const int exponent) noexcept;              \
-  template tnsr::a<DTYPE(data), DIM(data), FRAME(data)>                       \
-  DampedHarmonicGauge_detail::                                                \
-      spacetime_deriv_of_power_log_factor_metric_lapse(                       \
-          const Scalar<DTYPE(data)>& lapse,                                   \
-          const tnsr::I<DTYPE(data), DIM(data), FRAME(data)>& shift,          \
-          const tnsr::A<DTYPE(data), DIM(data), FRAME(data)>&                 \
-              spacetime_unit_normal,                                          \
-          const tnsr::II<DTYPE(data), DIM(data), FRAME(data)>&                \
-              inverse_spatial_metric,                                         \
-          const Scalar<DTYPE(data)>& sqrt_det_spatial_metric,                 \
-          const tnsr::ii<DTYPE(data), DIM(data), FRAME(data)>&                \
-              dt_spatial_metric,                                              \
-          const tnsr::aa<DTYPE(data), DIM(data), FRAME(data)>& pi,            \
-          const tnsr::iaa<DTYPE(data), DIM(data), FRAME(data)>& phi,          \
-          const double g_exponent, const int exponent) noexcept;
-
 #define INSTANTIATE_DV_FUNC(_, data)                                           \
   template void damped_harmonic_h(                                             \
       const gsl::not_null<                                                     \
@@ -831,33 +530,14 @@ void spacetime_deriv_damped_harmonic_h(
       const double sigma_t_L2, const double t_start_S, const double sigma_t_S, \
       const double sigma_r) noexcept;
 
-#define INSTANTIATE_SCALAR_FUNC(_, data)                             \
-  template void DampedHarmonicGauge_detail::log_factor_metric_lapse( \
-      const gsl::not_null<Scalar<DTYPE_SCAL(data)>*> logfac,         \
-      const Scalar<DTYPE_SCAL(data)>& lapse,                         \
-      const Scalar<DTYPE_SCAL(data)>& sqrt_det_spatial_metric,       \
-      const double exponent) noexcept;                               \
-  template Scalar<DTYPE_SCAL(data)>                                  \
-  DampedHarmonicGauge_detail::log_factor_metric_lapse(               \
-      const Scalar<DTYPE_SCAL(data)>& lapse,                         \
-      const Scalar<DTYPE_SCAL(data)>& sqrt_det_spatial_metric,       \
-      const double exponent) noexcept;
-
-GENERATE_INSTANTIATIONS(INSTANTIATE, (1, 2, 3), (double, DataVector),
-                        (Frame::Inertial))
-
 GENERATE_INSTANTIATIONS(INSTANTIATE_DV_FUNC, (1, 2, 3), (DataVector),
                         (Frame::Inertial))
-
-GENERATE_INSTANTIATIONS(INSTANTIATE_SCALAR_FUNC, (double, DataVector))
 
 #undef DIM
 #undef DTYPE
 #undef FRAME
 #undef DTYPE_SCAL
-#undef INSTANTIATE
 #undef INSTANTIATE_DV_FUNC
 #undef INSTANTIATE_SCALAR_FUNC
 /// \endcond
-}  // namespace gauges
-}  // namespace GeneralizedHarmonic
+}  // namespace GeneralizedHarmonic::gauges
