+++ tests/Unit/Evolution/Systems/GeneralizedHarmonic/GaugeSourceFunctions/Test_DampedHarmonicGaugeQuantities.cpp
@@ -26,6 +26,7 @@
 #include "Domain/Mesh.hpp"
 #include "Domain/Tags.hpp"
 #include "Evolution/Systems/GeneralizedHarmonic/GaugeSourceFunctions/DampedHarmonic.hpp"
+#include "Evolution/Systems/GeneralizedHarmonic/GaugeSourceFunctions/DampedWaveHelpers.hpp"
 #include "Evolution/Systems/GeneralizedHarmonic/Tags.hpp"
 #include "Framework/CheckWithRandomValues.hpp"
 #include "Framework/SetupLocalPythonEnvironment.hpp"
@@ -54,22 +55,7 @@ struct dt;
 }  // namespace Tags
 /// \endcond
 
-namespace GeneralizedHarmonic {
-namespace gauges {
-namespace DampedHarmonicGauge_detail {
-// The `detail` functions below are forward-declared to enable their independent
-// testing
-template <size_t SpatialDim, typename Frame, typename DataType>
-void spatial_weight_function(gsl::not_null<Scalar<DataType>*> weight,
-                             const tnsr::I<DataType, SpatialDim, Frame>& coords,
-                             double sigma_r) noexcept;
-
-template <size_t SpatialDim, typename Frame, typename DataType>
-void spacetime_deriv_of_spatial_weight_function(
-    gsl::not_null<tnsr::a<DataType, SpatialDim, Frame>*> d4_weight,
-    const tnsr::I<DataType, SpatialDim, Frame>& coords, double sigma_r,
-    const Scalar<DataType>& spatial_weight_function) noexcept;
-
+namespace GeneralizedHarmonic::gauges::DampedHarmonicGauge_detail {
 // The return-by-value implementations of spatial_weight_function and
 // spacetime_deriv_of_spatial_weight_function are intentionally only available
 // in the test because while convenient the additional allocations are bad for
@@ -99,77 +85,7 @@ double roll_on_function(double time, double t_start, double sigma_t) noexcept;
 
 double time_deriv_of_roll_on_function(double time, double t_start,
                                       double sigma_t) noexcept;
-
-template <typename DataType>
-Scalar<DataType> log_factor_metric_lapse(
-    const Scalar<DataType>& lapse,
-    const Scalar<DataType>& sqrt_det_spatial_metric, double exponent) noexcept;
-
-template <size_t SpatialDim, typename Frame, typename DataType>
-tnsr::a<DataType, SpatialDim, Frame> spacetime_deriv_of_log_factor_metric_lapse(
-    const Scalar<DataType>& lapse,
-    const tnsr::I<DataType, SpatialDim, Frame>& shift,
-    const tnsr::A<DataType, SpatialDim, Frame>& spacetime_unit_normal,
-    const tnsr::II<DataType, SpatialDim, Frame>& inverse_spatial_metric,
-    const Scalar<DataType>& sqrt_det_spatial_metric,
-    const tnsr::ii<DataType, SpatialDim, Frame>& dt_spatial_metric,
-    const tnsr::aa<DataType, SpatialDim, Frame>& pi,
-    const tnsr::iaa<DataType, SpatialDim, Frame>& phi,
-    double exponent) noexcept;
-
-template <size_t SpatialDim, typename Frame, typename DataType>
-void spacetime_deriv_of_power_log_factor_metric_lapse(
-    gsl::not_null<tnsr::a<DataType, SpatialDim, Frame>*> d4_powlogfac,
-    const Scalar<DataType>& lapse,
-    const tnsr::I<DataType, SpatialDim, Frame>& shift,
-    const tnsr::A<DataType, SpatialDim, Frame>& spacetime_unit_normal,
-    const tnsr::II<DataType, SpatialDim, Frame>& inverse_spatial_metric,
-    const Scalar<DataType>& sqrt_det_spatial_metric,
-    const tnsr::ii<DataType, SpatialDim, Frame>& dt_spatial_metric,
-    const tnsr::aa<DataType, SpatialDim, Frame>& pi,
-    const tnsr::iaa<DataType, SpatialDim, Frame>& phi, double g_exponent,
-    int exponent) noexcept;
-
-template <size_t SpatialDim, typename Frame, typename DataType>
-tnsr::a<DataType, SpatialDim, Frame>
-spacetime_deriv_of_power_log_factor_metric_lapse(
-    const Scalar<DataType>& lapse,
-    const tnsr::I<DataType, SpatialDim, Frame>& shift,
-    const tnsr::A<DataType, SpatialDim, Frame>& spacetime_unit_normal,
-    const tnsr::II<DataType, SpatialDim, Frame>& inverse_spatial_metric,
-    const Scalar<DataType>& sqrt_det_spatial_metric,
-    const tnsr::ii<DataType, SpatialDim, Frame>& dt_spatial_metric,
-    const tnsr::aa<DataType, SpatialDim, Frame>& pi,
-    const tnsr::iaa<DataType, SpatialDim, Frame>& phi, double g_exponent,
-    int exponent) noexcept;
-}  // namespace DampedHarmonicGauge_detail
-}  // namespace gauges
-}  // namespace GeneralizedHarmonic
-
-// Wrap `wrap_spacetime_deriv_of_power_log_factor_metric_lapse` here to make its
-// last argument a double, allowing for `pypp::check_with_random_values` to
-// work.
-template <size_t SpatialDim, typename Frame, typename DataType>
-tnsr::a<DataType, SpatialDim, Frame>
-wrap_spacetime_deriv_of_power_log_factor_metric_lapse(
-    const Scalar<DataType>& lapse,
-    const tnsr::I<DataType, SpatialDim, Frame>& shift,
-    const tnsr::A<DataType, SpatialDim, Frame>& spacetime_unit_normal,
-    const tnsr::II<DataType, SpatialDim, Frame>& inverse_spatial_metric,
-    const Scalar<DataType>& sqrt_det_spatial_metric,
-    const tnsr::ii<DataType, SpatialDim, Frame>& dt_spatial_metric,
-    const tnsr::aa<DataType, SpatialDim, Frame>& pi,
-    const tnsr::iaa<DataType, SpatialDim, Frame>& phi, const double g_exponent,
-    const double d_exponent) noexcept {
-  const auto exponent = static_cast<int>(d_exponent);
-  tnsr::a<DataType, SpatialDim, Frame> d4_powlogfac{};
-  GeneralizedHarmonic::gauges::DampedHarmonicGauge_detail::
-      spacetime_deriv_of_power_log_factor_metric_lapse(
-          make_not_null(&d4_powlogfac), lapse, shift, spacetime_unit_normal,
-          inverse_spatial_metric, sqrt_det_spatial_metric, dt_spatial_metric,
-          pi, phi, g_exponent, exponent);
-  return d4_powlogfac;
-}
+}  // namespace GeneralizedHarmonic::gauges::DampedHarmonicGauge_detail
 
 namespace {
 using Affine = domain::CoordinateMaps::Affine;
@@ -202,29 +118,6 @@ void test_options() noexcept {
 
 template <size_t SpatialDim, typename Frame, typename DataType>
 void test_detail_functions(const DataType& used_for_size) noexcept {
-  // weight_function
-  pypp::check_with_random_values<2>(
-      static_cast<Scalar<DataType> (*)(
-          const tnsr::I<DataType, SpatialDim, Frame>&, const double)>(
-          &::GeneralizedHarmonic::gauges::DampedHarmonicGauge_detail::
-              spatial_weight_function<SpatialDim, Frame, DataType>),
-      "Evolution.Systems.GeneralizedHarmonic.GaugeSourceFunctions."
-      "DampedHarmonic",
-      "spatial_weight_function",
-      {{{-10., 10.}, {std::numeric_limits<double>::denorm_min(), 10.}}},
-      used_for_size);
-  // spacetime_deriv_of_spatial_weight_function
-  pypp::check_with_random_values<2>(
-      static_cast<tnsr::a<DataType, SpatialDim, Frame> (*)(
-          const tnsr::I<DataType, SpatialDim, Frame>&, const double)>(
-          &::GeneralizedHarmonic::gauges::DampedHarmonicGauge_detail::
-              spacetime_deriv_of_spatial_weight_function<SpatialDim, Frame,
-                                                         DataType>),
-      "Evolution.Systems.GeneralizedHarmonic.GaugeSourceFunctions."
-      "DampedHarmonic",
-      "spacetime_deriv_spatial_weight_function",
-      {{{-10., 10.}, {std::numeric_limits<double>::denorm_min(), 10.}}},
-      used_for_size);
   // roll_on_function
   pypp::check_with_random_values<1>(
       &::GeneralizedHarmonic::gauges::DampedHarmonicGauge_detail::
@@ -241,42 +134,6 @@ void test_detail_functions(const DataType& used_for_size) noexcept {
       "DampedHarmonic",
       "time_deriv_roll_on_function",
       {{{std::numeric_limits<double>::denorm_min(), 10.}}}, used_for_size);
-  // log_factor_metric_lapse
-  pypp::check_with_random_values<1>(
-      static_cast<Scalar<DataType> (*)(const Scalar<DataType>&,
-                                       const Scalar<DataType>&, const double)>(
-          &::GeneralizedHarmonic::gauges::DampedHarmonicGauge_detail::
-              log_factor_metric_lapse<DataType>),
-      "Evolution.Systems.GeneralizedHarmonic.GaugeSourceFunctions."
-      "DampedHarmonic",
-      "log_fac", {{{std::numeric_limits<double>::denorm_min(), 10.}}},
-      used_for_size);
-  // spacetime_deriv_of_log_factor_metric_lapse
-  pypp::check_with_random_values<1>(
-      static_cast<tnsr::a<DataType, SpatialDim, Frame> (*)(
-          const Scalar<DataType>&, const tnsr::I<DataType, SpatialDim, Frame>&,
-          const tnsr::A<DataType, SpatialDim, Frame>&,
-          const tnsr::II<DataType, SpatialDim, Frame>&, const Scalar<DataType>&,
-          const tnsr::ii<DataType, SpatialDim, Frame>&,
-          const tnsr::aa<DataType, SpatialDim, Frame>&,
-          const tnsr::iaa<DataType, SpatialDim, Frame>&, const double)>(
-          &::GeneralizedHarmonic::gauges::DampedHarmonicGauge_detail::
-              spacetime_deriv_of_log_factor_metric_lapse<SpatialDim, Frame,
-                                                         DataType>),
-      "Evolution.Systems.GeneralizedHarmonic.GaugeSourceFunctions."
-      "DampedHarmonic",
-      "spacetime_deriv_log_fac",
-      {{{std::numeric_limits<double>::denorm_min(), 10.}}}, used_for_size,
-      1.0e-11);
-  // spacetime_deriv_of_power_log_factor_metric_lapse
-  pypp::check_with_random_values<1>(
-      &wrap_spacetime_deriv_of_power_log_factor_metric_lapse<SpatialDim, Frame,
-                                                             DataType>,
-      "Evolution.Systems.GeneralizedHarmonic.GaugeSourceFunctions."
-      "DampedHarmonic",
-      "spacetime_deriv_pow_log_fac",
-      {{{std::numeric_limits<double>::denorm_min(), 10.}}}, used_for_size,
-      1.e-11);
 }
 
 //
